Change structure of files
