+++ message/jobs/setOnQueue.js
@@ -1,25 +1,16 @@
 require("dotenv").config();
 const Queue = require("bull");
-const uniqid = require("uniqid");
 
-const createMessage = require("../database/createMessage");
-const transactionUpdate = require("../controllers/transactionUpdate");
-const breaker = require("../controllers/breaker/breaker");
+const transactionUpdate = require("../database/transactionUpdate");
+const breaker = require("../clients/breaker/breaker");
 const { REDIS_PORT } = process.env;
 
 const messageQueue = new Queue("charged-queue", `redis://${REDIS_PORT}`);
 const creditQueue = new Queue("credit-queue", `redis://${REDIS_PORT}`);
 const creditBackQueue = new Queue("creditBack-queue", `redis://${REDIS_PORT}`);
+
 const { MAX_JOBS, GOOD_JOBS } = process.env;
 let openQueue = true;
-
-const saveOnDatabase = (_id, destination, body) => {
-  return createMessage("primary", _id, destination, body, "pending").then(
-    () => {
-      return createMessage("replica", _id, destination, body, "pending");
-    }
-  );
-};
 
 const errorCredit = data => {
   const { _id } = data.message;
@@ -52,11 +43,11 @@ messageQueue.process((job, done) => {
 });
 
 const countingJobs = queue => {
-  queue.count().then(num => {
+  return queue.count().then(num => {
     if (num >= MAX_JOBS) {
       console.log("Queue busy");
       openQueue = false;
-    } else if (num <= GOOD_JOBS) {
+    } else if (num <= GOOD_JOBS && openQueue == false) {
       console.log("Back to paradise");
       openQueue = true;
     }
@@ -64,21 +55,14 @@ const countingJobs = queue => {
 };
 
 const setOnQueue = (_id, destination, body) => {
-  countingJobs(creditBackQueue);
-  if (openQueue) {
-    creditQueue.add({ _id, destination, body });
-  } else {
-    transactionUpdate(_id, "Message is busy, come back later");
-  }
-};
-
-const startQueue = (req, res) => {
-  const { destination, body } = req.body;
-  const _id = uniqid();
-  saveOnDatabase(_id, destination, body).then(() => {
-    setOnQueue(_id, destination, body);
-    res.send({ messageID: _id });
+  countingJobs(creditBackQueue).then(() => {
+    if (openQueue) {
+      creditQueue.add({ _id, destination, body });
+    } else {
+      transactionUpdate(_id, "Message is busy, come back later");
+    }
   });
 };
 
-module.exports = startQueue;
+
+module.exports = setOnQueue;
